refactor slider as a component and created dev data
