+++ components/ui/label.tsx
@@ -1,6 +1,13 @@
 import { cn } from "@/lib/utils";
 
-export const Label = ({ children, className }: { children: React.ReactNode, className?: string }) => {
+/**
+ * A simple label component.
+ *
+ * @param {React.ReactNode} children - The children element(s) to render within the label.
+ * @param {string} [className] - The class name to apply to the label element.
+ * @returns {JSX.Element}
+ */
+const Label = ({ children, className }: { children: React.ReactNode, className?: string }): JSX.Element => {
   return (
     <label
       className={cn(
@@ -12,3 +19,5 @@ export const Label = ({ children, className }: { children: React.ReactNode, clas
     </label>
   );
 };
+
+export default Label;
