+++ dev-data/classes.js
@@ -1,0 +1,11 @@
+import Biology from "../public/images/static-images/biology.webp";
+import History from "../public/images/static-images/history.webp";
+
+const classes = [
+  { id: '1', name: "biology", image: Biology},
+  { id: '2', name: "chemistry", image: History},
+  { id: '3', name: "biology", image: Biology},
+  { id: '4', name: "chemistry", image: History},
+];
+
+export default classes;
